# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "My Cat and dog server api swagger",
      "description": "Yet Another Node.js Cat and dog server api swagger",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths":{
        "/cat":{
            "get" :{
                "tags":["Cats"],
                "summary": "Get all Cats",
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Cats"
                        }
                    }
                }
            },
            "post":{
                "tags":["Cats"],
                "summary": "Create a new  Cat",
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"The cat we want to add ",
                        "schema":{
                            "$ref": "#definitions/Cat"
                        }
                    }
                ],
                "products":["application/json"],
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Cats"
                        }
                    }
                }
            }
        },
        "/cat/{id}":{
            "parameters":[
                {
                    "in":"path",
                    "name":"id",
                    "required":true,
                    "description":"ID of the cat that we want to match ",
                    "type":"string"
                }
            ],
            "get": {
                "tags":["Cats"],
                "summery": "get cat with given id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Cat with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ], 
                
                "products":["application/json"],
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Cat"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update cat with given ID",
                "tags": [
                    "Cats"
                ],
                "requestBody": {
                    "description": "Cat Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Cat"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Cat with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Cat with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/Cat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Cat"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    },
                    "404": {
                        "description": "Failed. Cat not found."
                    }
                }
            },
            "delete": {
                "summary": "Delete cat with given ID",
                "tags": [
                    "Cats"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete Cat with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "404": {
                        "description": "Failed. Cat not found."
                    }
                }
            }
        },
        "/cats":{
            
            "get": {
                "tags":["Cats"],
                "summary": "get cats with given search",
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"The cats we want to find ",
                        "schema":{
                            "$ref": "#definitions/Cat"
                        }
                    }
                ], 
                
                "products":["application/json"],
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Cats"
                        }
                    }
                }
            }
        },
...
        "/dog":{
            "get" :{
                "tags":["Dogs"],
                "summary": "Get all Dods",
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Dogs"
                        }
                    }
                }
            },
            "post":{
                "tags":["Dogs"],
                "summary": "Create a new  Dog",
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"The cat we want to add ",
                        "schema":{
                            "$ref": "#definitions/Dog"
                        }
                    }
                ],
                "products":["application/json"],
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Dogs"
                        }
                    }
                }
            }
        },
        "/dog/{id}":{
            "parameters":[
                {
                    "in":"path",
                    "name":"id",
                    "required":true,
                    "description":"ID of the dog that we want to match ",
                    "type":"string"
                }
            ],
            "get": {
                "summary": "get dog with given id",
                "tags":["Dogs"],                
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Dog with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ], 
                
                "products":["application/json"],
                "responses":{
                    "200":{
                        "description":"ok",
                        "schima":{
                            "$ref": "#definitions/Dog"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Dog with given ID",
                "tags": [
                    "Dogs"
                ],
                "requestBody": {
                    "description": "Dog Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Dog"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Dog with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "Dog with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/Dog"
                        }
                    }
                    
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Dog"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    },
                    "404": {
                        "description": "Failed. Dog not found."
                    }
                }
            },
            "delete": {
                "summary": "Delete Dog with given ID",
                "tags": [
                    "Dogs"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete Cat with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "404": {
                        "description": "Failed. Dog not found."
                    }
                }
            }
        }
    },
    "definitions":{
        "id":{
            "properties":{
                "uuid":{
                    "type": "string"
                }
            }
        },
        "Cat":{
            "type" : "object",
            "properties":{
                "name": { "type": "string"},
                "age": {"type" : "integer"},
                "type": { "type": "string"}
            }
        },
...
        "Dog":{
            "type" : "object",
            "properties":{
                "name": { "type": "string"},
                "age": {"type" : "integer"},
                "type": { "type": "string"},
                "color": { "type": "string"}
            }
        },
...
    }
}